Applying PR feedback
